product-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
